% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaNER.R
\docType{data}
\name{germaner_get_named_entities}
\alias{germaner_get_named_entities}
\alias{germaner_jar_filename}
\alias{germaner_download_jar}
\alias{germaparl_ne_regions}
\title{Get Named Entities using GermaNER}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 468 rows and 4 columns.}
\usage{
germaner_get_named_entities(x, mem = "-Xmx4g")

germaner_jar_filename(filename = "GermaNER-09-09-2015.jar")

germaner_download_jar()

germaparl_ne_regions
}
\arguments{
\item{x}{A \code{partition} object.}

\item{mem}{Java flags passed into system call.}

\item{filename}{Name of the GermaNER jar file.}
}
\value{
A \code{data.table}.
}
\description{
Use GermaNER (\url{https://github.com/tudarmstadt-lt/GermaNER}) for named
entity recognition.
}
\details{
The package includes a small data set called 'germaparl_ne_regions'
}
\examples{
\dontrun{
library(polmineR)
library(magrittr)
if (!file.exists(germaner_jar_filename())) germaner_download_jar()

x <- corpus("GERMAPARLMINI") \%>\% subset(speaker == "Angela Dorothea Merkel")
y <- germaner_get_named_entities(x = x)
}
}
\keyword{datasets}
